Rewrite to batch polling mode
This includes the top 3 countries rewritten for batch mode.

The first work on getting the flights per hour above the Netherlands.
For now this has mocked filter function, which is in the Observable
monad to make it easier to replace with the real service call.

Though the service call might be a batch thing also.
And it has a few tests for working with the JavaScript date api.
and grouping flights by their hours to record them.