Share hot observable flight data across components
page owns the observable and turns it hot it then passes it down to
child components as props.

When a component subscribes they immediately get the lastest flight
data, afterwards they recieve polling updates.
The subscription is done in the useEffect hook, which only depends on
the setData value, meaning it should only run once per component