Finish the geo filter api
This includes valdation for all the data using super-struct. Casting
from any became a mess. It fixes the latitude and longitude bug for the
geolocation. Includes new test data, with a plane I knew about that was
above the Netherlands for debuging and testing purposes.
The number in the current UI is really high because of the
interpolation, this is weird since planes stay above the Netherlands for
a long time. I need to rethink that feature and rethink the flights
since application started running feature